feat: add route registration hook for custom/premium packages
Custom packages can now register API routes without editing the
framework-owned router.ts. Users add handlers in the user-zone
file `ottabase/config.routes.ts` (same pattern as config.migrations.ts).

- Extract ApiRouteContext to worker/routes/types.ts (shared type)
- Add ottabase.template/config.routes.ts with scaffold + docs
- Wire handleCustomRoutes() into resolveApiRoute() as fallback
  after all built-in routes
- Update README with route registration in premium package workflow
